pyproject.toml
### Build ###
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = ["nebari_workflow_controller"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "nebari_workflow_controller/_version.py"
local_scheme = "node-and-timestamp"

# [project.entry-points."setuptools_scm.version_scheme"]
# "calver-by-date" = "setuptools_scm.version:calver_by_date"

### Project ###
[project]
name = "nebari-workflow-controller"
dynamic = ["version"]
description = "An admission controller for argo workflows in Nebari"
readme = "README.md"
requires-python = ">=3.8"
license = "BSD-3-Clause"
authors = [
    { name = "Nebari development team", email = "internal-it@quansight.com" }
]
keywords = [
    "nebari",
    "argo-workflows",
    "kubernetes",
    "fastapi",
    "admission-controller",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
]
dependencies = [
    "python-keycloak==2.15.3",
    "fastapi==0.95.1",
    "uvicorn==0.21.1",
    "pydantic==1.10.7",
    "jsonpatch==1.32",
]

[project.optional-dependencies]
dev = [
    "black==22.3.0",
    "flake8==3.8.4",
    "pre-commit",
    "pytest",
    "pytest-mock",
    "pyyaml",
    "build",
]

[project.urls]
Documentation = "https://github.com/nebari-dev/nebari-workflow-controller"
Source = "https://github.com/nebari-dev/nebari-workflow-controller"

[tool.ruff]
ignore = [
    "E501", # Line too long
    "F821", # Undefined name
]
extend-exclude = [
    "__pycache__"
]
